Add files via upload
Excited to share my first-ever ML project where I tried to build a simple linear regression model in Python! With this model, I was able to predict the relationship between the platform used for marketing and the sales generated through that platform and identify any potential correlations. I implemented this machine learning algorithm from scratch. It was an incredible learning experience, and I'm excited to share my results with you all. Looking forward to taking on more projects like this in the future and expanding my skills in data science.