Add files via upload
I just built a simple linear regression model to predict salaries based on years of experience. The model achieved an impressive R-squared value of .8989, indicating a strong correlation between these two variables. The steps include exploring the dataset, cleaning and preprocessing the data, and then training and evaluating the model. Feedback and suggestions are always welcome!